% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrelrates.R
\name{allnodedist2}
\alias{allnodedist2}
\title{Get the distance between 2 nodes / tips.}
\usage{
allnodedist2(node1, node2, t)
}
\arguments{
\item{node1}{a tip or node}

\item{node2}{a tip or node}

\item{t}{tree containing the nodes}
}
\description{
Get the distance between 2 nodes / tips.
}

