% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrelrates.R
\name{decendDist}
\alias{decendDist}
\title{Get the relative rates of two decendant branches.}
\usage{
decendDist(bnode, t, finalrates)
}
\arguments{
\item{bnode}{a node in the tree with two decendants}

\item{t}{the tree}

\item{finalrates}{ancestral relative rates, which are used as multipliers
Method:
Get the average distance from bnode to all decendant tips (avgdist).
For each of bnode's direct child nodes (c1,c2):
Get the total distance from this node to each tip (c1dist,c2dist).
Divide this total by avgdist (rate1,rate2).
If bnode is not the root node:
Multiply each rate by the ancestor relative rate.}
}
\description{
Get the relative rates of two decendant branches.
}
\references{
Tamura et al. 2012 PNAS
}

