% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrelrates.R
\name{geneinfo}
\alias{geneinfo}
\title{Get the relative rates on a tree.}
\usage{
geneinfo(t)
}
\arguments{
\item{t}{tree
Each rate is for a branch, which is labeled by the two nodes it connects}
}
\description{
Get the relative rates on a tree.
}
\references{
Tamura et al. 2012 PNAS
}

