% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rreltrates.R
\name{outsideCI}
\alias{outsideCI}
\title{Checks if value is outside confidence interval.}
\usage{
outsideCI(rate, lb, ub)
}
\description{
Takes a value, upper bound (mean+2sd), and lower bound (mean-2sd)
}

