% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rreltrates.R
\name{geneinfo}
\alias{geneinfo}
\title{Get the mean and var of relative rates on a tree.
Also returns branches with rates outside the CI}
\usage{
geneinfo(gene)
}
\description{
Takes a gene name.
Gets the primate tree for this gene using rensembl.
Gets all relative rates.
}

