% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrelrates.R
\name{outsideCI}
\alias{outsideCI}
\title{Checks if value is outside confidence interval.}
\usage{
outsideCI(rate, lb, ub)
}
\arguments{
\item{rate}{a value of interest}

\item{lb}{lower bound of confidence interval}

\item{upper}{bound of confidence interval}
}
\description{
Checks if value is outside confidence interval.
}

